% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensus2013.R
\docType{data}
\name{nzcensus2013}
\alias{nzcensus2013}
\title{New Zealand 2013 census data}
\format{A data frame with 416,085 rows and 11 variables:
\itemize{
\item \code{sheet} the name of the table and the worksheet that the data was imported
from
\item \code{title} the title of the table
\item \code{dimensions} the subgroups that the table uses
\item \code{population} the population that the table reports
\item \code{row} the row number of the cell that the data was imported from
\item \code{col} the column number of the cell that the data was imported from
\item \code{value} the number of people observed in the given subset of the New
Zealand population
\item \code{col_header1} a type of subset of the New Zealand population
\item \code{col_header2} a subset of the New Zealand population
\item \code{col_header3} a subset of the New Zealand population
\item \code{row_header1} a type of subset of the New Zealand population
\item \code{row_header2} a subset of the New Zealand population
\item \code{row_header3} a subset of the New Zealand population
\item \code{row_header4} a subset of the New Zealand population
}}
\source{
http://www.stats.govt.nz/Census/2013-census/data-tables/electorate-tables.aspx
released under the CC BY 4.0 licence
https://creativecommons.org/licenses/by/4.0/
}
\usage{
nzcensus2013
}
\description{
A dataset containing aggregate statistics from the 2013 New Zealand census.
}
\details{
This data has been randomly rounded to protect confidentiality.  Individual
figures may not add up to totals, and values for the same data may vary in
different tables.

Source data symbols "..C" for "confidential" and "..." for "not applicable"
have both been imported as \code{NA}

Refer to the source spreadsheet for other notes.
}
\keyword{datasets}
